+++ app/src/main/java/com/example/mattshealthtracker/CorrelationRepository.kt
@@ -238,7 +238,7 @@ class CorrelationRepository(context: Context) {
                             symptomB = it.getString(symptomBIndex),
                             lag = it.getInt(lagIndex),
                             isPositiveCorrelation = it.getInt(isPositiveIndex) == 1,
-                            confidence = it.getDouble(confidenceIndex),
+                            confidence = it.getFloat(confidenceIndex),
                             insightfulnessScore = it.getDouble(insightfulnessIndex),
                             preferenceScore = it.getInt(preferenceIndex),
                             lastCalculatedDate = it.getLong(lastCalculatedIndex)
